+++ src/main/kotlin/com/zwolsman/bombastic/domain/converters/TileReadingConverter.kt
@@ -12,7 +12,7 @@ object TileReadingConverter : Converter<String, Tile> {
         val id = source.substring(0 until 2).toInt()
         return when (source[2]) {
             'B' -> Bomb(id, revealedByUser = source[3] == 'T')
-            'P' -> Points(id, amount = source.substring(startIndex = 3).toInt())
+            'P' -> Points(id, amount = source.substring(startIndex = 3).toLong())
             else -> throw Exception("Invalid tile state")
         }
     }
